requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.3.1
django-allauth-ui==1.5.3
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
slippers==0.6.2
sqlparse==0.5.3
stripe==11.4.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
cloudinary